moving + firehose
